requirements-test.txt
pytest-cov>=2.5.1
pytest-mock>=1.7.0
